+++ src/type/ft_typeX.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_typeX.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hherin <hherin@student.42.fr>              +#+  +:+       +#+        */
+/*   By: lucaslefrancq <lucaslefrancq@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/17 16:24:40 by hherin            #+#    #+#             */
-/*   Updated: 2019/12/18 16:34:18 by hherin           ###   ########.fr       */
+/*   Updated: 2021/11/17 15:43:44 by lucaslefran      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ static char	*ft_zeropad(char *s, char *tmp, s_option opt, size_t size)
 	}
 	else
 	{
-		abso = (opt.width < 0) ? ft_strlen(s) : opt.width;
+		abso = (opt.width < 0) ? (int)ft_strlen(s) : opt.width;
 		zero = (opt.flag.zero) ? abso - ft_strlen(s): 0;
 		start = (opt.flag.minus || opt.width < 0) ? 0 : size - FT_MAX(ft_strlen(s), zero);
 		while (i++ < zero)
